Restructure scripts code
