+++ scripts/lib/aws-client.js
