+++ scripts/package.json
@@ -21,6 +21,6 @@
   "scripts": {
     "ses": "aws-ses-local -o ./out",
     "dynamodb": "node node_modules/local-dynamo/bin/launch_local_dynamo.js --port 4567",
-    "lint": "eslint src/ tools/"
+    "lint": "eslint lib/ tools/"
   }
 }
