+++ scripts/tools/run-thunder-command.js
